requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
APScheduler>=3.10.4
pandas>=2.1.0
